requirements.txt
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
nltk==3.8.1
openai==1.14.2
poppler-utils==0.1.0
pymilvus==2.4.0
pypdf==4.2.0
python-dotenv==1.0.1
tiktoken==0.6.0
llama-index-embeddings-openai
llama-index-readers-file
llama-index-embeddings-langchain
llama-index-llms-openai
sentence-transformers
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.2
pdfrw==0.4
pyside6==6.6.1 
pymupdf==1.24.7
tiktoken